Add AssignParameterOffset.
This module allows you to, from either the calibration or estimation frameworks, assign a floating point value using the combination of a base value and an offset to multiple parameters.  Each parameter can have a different base value.  This is useful when calibrating the Passenger and Carpool modes which have significant overlap.