Fix reflecting on unloaded model systems
Currently if you are running XTMF with a debugger you can end up in cases where the configuration will have projects
with additional model systems stored in the project but there is no module loaded for that model system.  With
delayed loading that will be filled with a null and crash the model run.  This patch insteads has the reflection
skip over model systems that are not loaded.