Update modeller bridge to be more resilient in the case a tool is not found. (#316)
If a tool is not found the modeller bridge will not finish reading in all of the parameters from the C# side of the bridge, leading to invalid codes being sent through the bridge.  This patch makes sure to flush all of that content before doing the check and potentially sending the error.