Fix race condition in location choice estimation
In the current implementation if you have EstimationMode set the three location choice models will load in parallel.  This
however causes a race condition for the temporary storage stored for each time period.  This patch checks to see if
we are going to be using estimation mode, and if so will load the models sequentially.