Fix white text on light background for parameter filters (#327)
In the current implementation when using the light mode the text for both the quick parameter and the module parameters is white.
In this patch we instead change it to use MaterDesignBody.  This will give black text when using a light theme and white
text when using the dark theme.