Multi-Langage Support (#332)
* Multi Language Support for floating point numbers

Currently XTMF normally uses the local culture when processing values, but this runs into issues when data might be coming
in from both the local culture and in the standard invariant culture.  These set of sweeping changes strives to
specify which culture to use where appropriate, and if we can't know default to trying the local culture first and then
falling back to the invariant culture as a second option.

* Updated the writing of CSV outputs to use the Invariant Culture to avoid cases where commas are unexpectedly being added.