Added peak hour factor and cascading demand for DAT station capacities.
In the previous implementations of ComputeStationCapacityFactor each time period was computed independently.  This version allows for the additional
demand to cascade down between time periods.  The order of the time periods is assumed to be the same as the submodule list.  This behaviour is off
by default to keep backwards compatibility.  An additional change is the ability to set the time period peak hour factors.  Previously demand
was assumed to be in peak period.  This option gives more flexibility going forward.  The PHF will be divided back out in order to get the
period volumes.